+++ dapps/shop/backend/package.json
@@ -12,7 +12,9 @@
   "dependencies": {
     "aws-sdk": "^2.572.0",
     "body-parser": "^1.19.0",
+    "bottleneck": "^2.19.5",
     "bs58": "^4.0.1",
+    "commander": "^4.1.1",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
