+++ dapps/shop/src/pages/Products.js
@@ -130,6 +130,11 @@ require('react-styl')(`
         .price
           font-weight: bold
           margin-top: 0.25rem
+          span.shipping
+            opacity: 0.6
+            font-weight: normal
+            margin-left: 0.5rem
+            font-size: 0.75rem
 
   @media (max-width: 767.98px)
     .products
